Added additional selection box
Users now have the ability to specify in greater detail what text should be kept in the renaming process.

Updated .gitignore

Added a TODO idea to add a temporary file backup folder which could allow users to revert bad renaming, as it can't be undone from File Explorer. Would add a dialog asking the user to check on the updated files, and if they click "Yes" or close out the window, the backup folder would be deleted.